Address PR review: deduplicate helpers, fix hints, tighten code
- Suppress find --hints (was emitting wrong Summary hints)
- Replace append.rs inline KV validation with set::parse_kv
- Simplify Status task filter to starts_with (zero-alloc)
- Extract duplicated format_iso8601 to commands/mod.rs
- Avoid cloning sequence in remove_value_in_memory
- Use method references instead of redundant closures
- Soften types.rs doc comment to reflect shared (not all) output shapes
